integration/aws/requirements.txt
# Python dependencies for lambda_function.py 
psycopg2-binary
crhelper